# Changes to Functions to include inline and moved benches to criterion to use stable toolchain