--- a/pnet_packet/Cargo.toml
+++ b/pnet_packet/Cargo.toml
@@ -16,5 +16,8 @@ pnet_base = { path = "../pnet_base", version = "0.29.0" }
 pnet_macros_support = { path = "../pnet_macros_support", version = "0.29.0" }
 pnet_macros = { path = "../pnet_macros", version = "0.29.0" }
 
+[dev-dependencies]
+hex = "0.4.3"
+
 [build-dependencies]
 glob = "0.3.0"