--- a/pnet_packet/Cargo.toml
+++ b/pnet_packet/Cargo.toml
@@ -18,6 +18,12 @@ pnet_macros = { path = "../pnet_macros", version = "0.29.0" }
 
 [dev-dependencies]
 hex = "0.4.3"
+criterion = {version = "0.3.5", features = ["html_reports"]} #added HTML feature becuase of the annoying warnings when running the tests
 
 [build-dependencies]
 glob = "0.3.0"
+
+[[bench]]
+name = "bench_pnet_packet"
+path = "benches/packet_benchmarks.rs"
+harness = false