pent_macro_support/src/packet.rs
- added inlining to all functions available to increase speed at cost of size

pnet_macro/src/decorator.rs
- added inlining to all functions to increase speed at cost of size

pnet_packet/
- moved benches from ethernet.rs to new folder ~/benches/packet_benchmarks.rs

Benchmarks Before Changes 
```
test bench_ipv4_parsing                  ... bench:           8 ns/iter (+/- 0)
test bench_packet_get_source             ... bench:          16 ns/iter (+/- 0)
test bench_packet_immutable_to_immutable ... bench:          10 ns/iter (+/- 1)
test bench_packet_mutable_to_immutable   ... bench:          10 ns/iter (+/- 0)
test bench_packet_new_constructor        ... bench:           8 ns/iter (+/- 0)
test bench_packet_set_source_black_box   ... bench:          18 ns/iter (+/- 1)
Benchmarks After Changes
```
```
running 6 tests
test bench_ipv4_parsing                  ... bench:           1 ns/iter (+/- 0)
test bench_packet_get_source             ... bench:           7 ns/iter (+/- 0)
test bench_packet_immutable_to_immutable ... bench:           2 ns/iter (+/- 0)
test bench_packet_mutable_to_immutable   ... bench:           2 ns/iter (+/- 0)
test bench_packet_new_constructor        ... bench:           1 ns/iter (+/- 0)
test bench_packet_set_source_black_box   ... bench:           9 ns/iter (+/- 0)
```