@infosechoudini 2022-04-19
Criterion Output

```
Gnuplot not found, using plotters backend
EthernetPacket New Packet                                                                             
                        time:   [1.9560 ns 1.9666 ns 1.9787 ns]
                        change: [-0.5771% +0.3666% +1.2536%] (p = 0.43 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

EthernetPacket Get Source                                                                             
                        time:   [4.5342 ns 4.5548 ns 4.5782 ns]
                        change: [+0.1741% +1.5730% +2.5907%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 13 outliers among 100 measurements (13.00%)
  8 (8.00%) high mild
  5 (5.00%) high severe

EthernetPacket Set Source                                                                             
                        time:   [637.33 ps 638.82 ps 640.36 ps]
                        change: [-3.0706% -2.4471% -1.8802%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

Mutable to Immutable    time:   [2.5357 ns 2.5486 ns 2.5620 ns]                                  
                        change: [-1.5676% -0.7663% -0.0239%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe

Immutable to Immutable  time:   [2.5807 ns 2.5930 ns 2.6076 ns]                                    
                        change: [-1.4937% -0.6850% +0.0496%] (p = 0.09 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe

IPV4 Parsing            time:   [1.9412 ns 1.9472 ns 1.9535 ns]                          
                        change: [+0.0893% +0.6524% +1.2177%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
```