@stappersg 2022-04-21
> @mrmonday et al., can i get a workflow approval?

The request is be seen by a human.
My advice is to ask again in about week.
Consider meanwhile the idea of something like _easter vacaction_ or any other possible reason for a delay in response.

My real message is
> @infosechoudini  aim for the long run, avoid to get worn down on the short run

Other thing
```text
Day changed to 20 apr 2022
00:22 -!- infosechoudini [~infosecho@h597.436.156.298.static.ip.redacted.net] has joined #libpnet
00:24 < infosechoudini> hey, i'm looking to help contribute to libpnet, was wondering what's the process
02:21 -!- infosechoudini [~infosecho@h597.436.156.298.static.ip.redacted.net] has quit [Quit: Connection closed]
```
_Feel welcome_ to hang-out at `libera/#libpnet` _IRC channel_.