@mrmonday 2022-04-25
I'm surprised the annotations make so much difference - in the past we've removed lots of them because the compiler is "sufficiently smart" that it made no difference.

It would be good to get the benchmarks in CI, no need to do that here though.

Thanks!